Intial commit of yomlette.
This design defers the lexing of template actions until the parsing of
the YAML itself in order to simplify the interactions between the two
parsers. From the point of view of the YAML parser, a template action is
opaque, and is treated as a single token. Inside the template parser,
the design adds a new item type, `itemYaml`, that encapsulates a YAML
fragment. This item takes the place of `itemText`.